+++ cmeta/utils/common.py
@@ -497,3 +497,21 @@ def sort_versions(versions, reverse=False):
         return (*nums, suffix)
 
     return sorted(versions, key=parse_version, reverse=reverse)
+
+###################################################################################################
+def flatten_dict(d, parent_key="", sep="."):
+    items = []
+    for k, v in d.items():
+        new_key = f"{parent_key}{sep}{k}" if parent_key else k
+
+        if isinstance(v, dict):
+            items.extend(flatten_dict(v, new_key, sep=sep).items())
+
+        elif isinstance(v, list):
+            # add comma before '=' by modifying the key
+            items.append((f"{new_key},", ",".join(map(str, v))))
+
+        else:
+            items.append((new_key, v))
+
+    return dict(items)